CRM-13244 - ActionScheduleTmp - Move entity-specific parts to separate classes
