+++ CRM/Core/BAO/ActionSchedule.php
@@ -92,7 +92,7 @@ class CRM_Core_BAO_ActionSchedule extends CRM_Core_DAO_ActionSchedule {
       $mappingId = $mapping->getId();
       $entityValueLabels[$mappingId] = $mapping->getValueLabels();
       // Not sure why: everything *except* contact-dates have a $valueLabel.
-      if ($mapping->getId() !== CRM_Core_ActionScheduleTmp::CONTACT_MAPPING_ID) {
+      if ($mapping->getId() !== CRM_Contact_ActionMapping::CONTACT_MAPPING_ID) {
         $valueLabel = array('- ' . strtolower($mapping->getValueHeader()) . ' -');
         $entityValueLabels[$mapping->getId()] = $valueLabel + $entityValueLabels[$mapping->getId()];
       }
