+++ CRM/Event/Page/ManageEvent.php
@@ -315,7 +315,7 @@ ORDER BY start_date desc
     );
     $this->assign('eventCartEnabled', $enableCart);
     $mapping = CRM_Utils_Array::first(CRM_Core_BAO_ActionSchedule::getMappings(array(
-      'id' => CRM_Core_ActionScheduleTmp::EVENT_NAME_MAPPING_ID,
+      'id' => CRM_Event_ActionMapping::EVENT_NAME_MAPPING_ID,
     )));
     $eventType = CRM_Core_OptionGroup::values('event_type');
     while ($dao->fetch()) {
